Inital pass at higher level VRP implementation
Accepts a list of locations and the number of vehicles in use and
calculates the appropriate number of routes.

With the location data supplied in `test/vrp_test.rb`, if I caluclate
the distance matrix and scale distances with `DISTANCE_SCALE` the
program segfaults (note, the sample locations data is "real world"
and represents several locations broadly within one city). If I use the
locations data from `test/tsp_test.rb` _without_ scaling distances, the
program segfaults. Seems that there is a point at which is makes sense
to scale distances, but untill that point is reached doing so causes
problems. I wonder if the `DISTANCE_SCALE` should be either dynamic
based on the distances between the points being routed, or user
configurable?

I've basically duplicated the distance matrix calculation from the TSP
solution. Could this be extracted into it's own class?

Similarly, right now TSP returns an instance of itself with four
`attr_accessor`s: `route`, `route_indexes`, `distances`,
`total_distance`. VRP currently simply returns an instance of itself
with the attr_accessor `routes`, which is an array of arrays containing
the routed locations. I feel like it would be cool if there were a
`Route` class that would tie up `route`, `route_indexes`, `distances`
and `total_distance`. TSP would then return a simple instance of the
`Route` class, VRP would return an array of `Route` classes. wdyt?

This is a real basic first pass. Please let me know if there are any
changes that should be made at this stage.

:v: