Explicitly add dimensions to VRP
First pass at making VRP more flexible. This takes the basic
implementation and breaks things up to allow injecting a "Dimension"
into the process via `add_dimension`. Currently this new method only
requires the name of the dimension to be given, and defaults the rest of
the required variables to the values given in the "Distance" basic VRP
example.

The basic VRP solution is now solved with

```ruby
ORTools::VRP.new(locations, 2).add_dimension("Distance").routes
```

Feels like this approach allows for building more flexible solutions at
the expence of complicating usage for the user. You now have to
understand a bit more about how Dimensions work in ORTools, and I expect
the logical conclusion of this path is that we eventually end up back at
the low-level API.

Instead, would we prefer to have a series of VRP classes teed up for
solving problems already, and that are only somewhat customisable? If
you want to go to town on a custom solution, the API is there, but if
you just want e.g. a solution that handles timeboxed deliveries, then
perhaps the `VRPTW` class is the class you want? Or is there a middle
ground that I haven't understood yet?